Upgrade to Squeekboard v1.16.0 (#14)
This PR also changes the way our Squeekboard patches are applied. Previously, a patch branch of Squeekboard was maintained on my personal Gnome Gitlab account. This made it difficult to keep track of all the moving parts. Now, Squeekboard patches go in a directory in this repo, which should be much easier to keep track of.

Patches 0001 and 0002 are identical to commits that were previously on my branch. 0003 is newly required by v1.16.0 (it fixes the source of the crash that made upgrading trouble).

Fixes #12.

Co-authored-by: William Wold <wm@wmww.sh>